+++ pci.h
@@ -17,7 +17,7 @@ typedef struct PCI_HM
 	struct PCI_LL *HashMap[HASHMAP_SIZE];
 } PCI_HM;
 
-int pci_lookup(PCI_HM *hashmap, uint64_t block_number);
+uint64_t pci_lookup(PCI_HM *hashmap, uint64_t block_number);
 void pci_insert(PCI_HM *hashmap, uint64_t block_number, uint64_t index);
 void pci_delete(PCI_HM *hashmap, uint64_t block_number);
